refractor: Move file and refractoring directory
